ParserSharing

Description

ParserSharing is a general parsing library which merges parse trees such that the input is consumed in one traversal. From the user perspective the parsers are expressed as classical recursive descendent parser using familiar parser combinators. The library is easily extensible, error control is also easily implementable and the parsing is already done in an online fashion. Marcelo Sousa, Utrecht University 2011.

Documentation

data Parser where

A Parser is defunctionalized with a GADT where the constructors represent the normal parsers combinators. The default token type is Char.

Constructors

Sym ∷ Char → Parser Char

Seq ∷ Parser (b → a) → Parser b → Parser a

Alt ∷ Parser a → Parser a → Parser a

Ret ∷ a → Parser a

data Pending where

Continuations are represented by Pending which in the simple case is either Done when there are no more parsers to process or a Stack of parsers. A Split pending means that the current parse tree was being shared and should be splitted in two. LazyShare represents that part in the future of the tree that will be shared. This is useful because of the associativity of Seq.

Constructors

Stack ∷ Parser p → Pending r → Pending (p, r)

Split ∷ Pending r1 → Pending r2 → Pending (Either r1 r2)

LazyShare ∷ Parser p → Pending r → Pending (Share p r)

Done ∷ Pending ()

data Share p r where

Share represents a pair with a parser and a continuation

Constructors

Share ∷ p → r → Share p r

type States a = [State a]

data State a

State is an existential type that contains a list of semantic function that receive a cartesian product of the results of the parsers and applies them according to their semantics.

Constructors

∀ b . State (Func b a) (Pending b)

data Func b a where

Generalizing Func for N alternatives we would need a list of functions.

Constructors

Sem ∷ [b → a] → Func b a

data Equal where

Equal data type for equality proof. The Eq construct expresses that two types are equal. This is particulary useful for equality when using existential types.

Constructors

Eq ∷ Equal a a

castParser ∷ Parser a → IO (StablePtr (Parser b))

castParser casts a parser to another type that is inferred to be able to compare Stable Pointers.

eq ∷ Parser a → Parser b → Maybe (Equal a b)

eq is the equality function between two parsers. Here a and b have the same type although the use of the existential types in State cannot allow us to assume so. That's the reason why the casting is safe. For comparison we use Stable Pointers.

runParser ∷ Parser a → [Char] → [a]

runParser is the main function which the user has to invoke.

parse ∷ States a → [Char] → [a]

parse is the iteratee that consumes input and updates the state by calling transition.

evalStates ∷ States a → [a]

evalStates triggers evaluation of the success functions. In that sense the process is not completely online. The functions is a bit more complicated than what it should be because of Ret.

transition ∷ States a → Char → States a

transition unfold and merges the states and then when their are ready to consume input it calls reduce.

reduce ∷ Char → States a → State a → States a

reduce consumes the token when the States are unfolded to a point where on top of the Stack we have a Sym or the pending is Done.

termination ∷ States a → Bool

termination check if all Pending in States are either Done or have a Sym on top of the Stack.

isFolded ∷ State a → Bool

unFoldAndMerge ∷ States a → States a

unFoldAndMerge unfolds and merges states until they are ready to consuming input. We start by unFolding the Split pending, then the LazyShare and then the State with Ret on top of Stack. This is because the sharing only happens with Seq and Alt which we process in the end. This is an highly inefficient function at the moment.

unFoldNtAtHeads ∷ States a → States a

unFoldNtAtHeads is a simple dispatch function that uses pattern match to call specific-unFolding functions.

unFoldSplitAtHead ∷ State a → States a

unFoldSplitAtHead processes a Split pending. Since when we are right-merging we simply take the head of the Sem list and apply Left on the result. The other State is the tail of Sem with Right applied to its result.

unFoldShareAtHeads ∷ States a → States a

unFoldShareAtHeads dispatches all the states with LazyShare to unFoldShareAtHead.

unFoldShareAtHead ∷ State a → States a → States a

unFoldShareAtHead is going to merge the pairs of LazyShare that have the same initial parser. It might happen that one of the element of the pair was already discarded so in that case it's simply converted into a Stack.

unFoldSeqAtHead ∷ State a → States a → (States a, States a)

unFoldSeqAtHead unfolds the constructors Seq with Seq, Seq with Ret and Seq with a generic top of a stack since all the others should be already unfolded by now. It compares both sequenced parsers for equality and applies Split and/or LazyShare depending if the their are equal.

unFoldRetAtHead ∷ State a → State a

unFoldRetAtHead simply puts the value of Ret on top of the stack of values.

unFoldAltAtHead ∷ State a → States a

unFoldAltAtHead traverses the Alt parser to get a list of alternatives unFoldAlt and uses nub to discard equal parsers to then build the list of corresponding States.

unFoldAlt ∷ Parser a → [Parser a]