ParserSharingCombinators

Description

ParserSharingCombinators defines some of the classic parser combinators using the ParserSharing library. Marcelo Sousa, Utrecht University 2011.

Documentation

(<|>) ∷ Parser a → Parser a → Parser a

Choice Parser Combinator

(<*>) ∷ Parser (b → a) → Parser b → Parser a

Sequence Parser Combinator

pSym ∷ Char → Parser Char

Token (Char) Parser Combinator

pReturn ∷ a → Parser a

Return Parser Combinator

(<$>) ∷ (b → a) → Parser b → Parser a

Semantic Parser Combinator

opt ∷ Parser a → a → Parser a

Optional Parser Combinator. Either the parser given succeds or the default value is returned.

pMany1 ∷ Parser a → Parser [a]

Multiple Parser Sequence Combinator.

pMany ∷ Parser a → Parser [a]