ParserSharingSimple

Description

ParserSharingSimple is a simple parsing library which was extending to allow sharing. I have published this library so that the user can understand the basic principle of how the library works without sharing and then to move to the shared version. Marcelo Sousa, Utrecht University 2011.

Documentation

data Parser where

A Parser is defunctionalized with a GADT where the constructors represent the normal parsers combinators. The default token type is Char. I've also added here the constructor Sat.

Constructors

Sym ∷ Char → Parser Char

Ret ∷ a → Parser a

Seq ∷ Parser (b → a) → Parser b → Parser a

Alt ∷ Parser a → Parser a → Parser a

Sat ∷ (Char → Bool) → Parser Char

data Pending where

Continuations are represented by Pending which in the simple case is either Done when there are no more parsers to process or a Stack of parsers.

Constructors

Stack ∷ Parser a → Pending b → Pending (a, b)

Done ∷ Pending ()

type States a = [State a]

data State a

State is an existential type that contains a semantic function that receive a cartesian product of the results of the parsers and applies them according to their semantics.

Constructors

∀ b . State (b → a) (Pending b)

runParser ∷ Parser a → [Char] → [a]

runParser is the main function which the user has to invoke.

parse ∷ States a → [Char] → [a]

parse is the iteratee that consumes input and updates the state by calling transition.

evalStates ∷ States a → [a]

evalStates triggers evaluation of the success functions. In that sense the process is not completely online. The functions is a bit more complicated than what it should be because of Ret. Notice that in case of failure the state is discarded.

transition ∷ States a → Char → States a

transition unfold the top of the stack until we have a Sym or 'Done and then calls reduce.

reduce ∷ Char → States a → State a → States a

reduce consumes the token when the States are unfolded to a point where on top of the Stack we have a Sym or the pending is Done.

unFoldNtAtHeads ∷ State a → States a

unFoldNtAtHeads pattern-matches on the top of the Pending and it's either a reduce state does not change it. If there is a Ret on the top of the Stack it will put the value on top of stack represented by the cartesian product and use that as argument to the semantic function. If there is a Seq on top of the Stack it will unfold it into a new Stack and it will apply the result function of the first parser to the result value of the second one and wrap that value into a pair that is the argument of the overall semantic function. If there is a Alt on top of the Stack it will simply create two states out of it. Note that unFoldNtAtHeads is being recursively called until we can't unFold anymore.